+++ mod-source-record-manager-server/src/test/java/org/folio/services/JournalServiceImplTest.java
@@ -42,7 +42,7 @@ public class JournalServiceImplTest {
   }
 
   @Test
-  public void shouldCalledDaoOnce() {
+  public void shouldCallSaveOnce() {
     JournalRecord journalRecord = new JournalRecord()
       .withJobExecutionId(UUID.randomUUID().toString())
       .withSourceId(UUID.randomUUID().toString())
@@ -64,7 +64,7 @@ public class JournalServiceImplTest {
   }
 
   @Test
-  public void shouldCalledManyTimes() {
+  public void shouldCallSaveBatchOnce() {
     JournalRecord journalRecord = new JournalRecord()
       .withJobExecutionId(UUID.randomUUID().toString())
       .withSourceId(UUID.randomUUID().toString())
@@ -83,10 +83,11 @@ public class JournalServiceImplTest {
       .add(jsonJournalRecord)
       .add(jsonJournalRecord);
 
-    when(journalRecordDao.save(any(JournalRecord.class), eq(TENANT_ID))).thenReturn(Future.succeededFuture());
+    when(journalRecordDao.saveBatch(any(JsonArray.class), eq(TENANT_ID))).thenReturn(Future.succeededFuture());
 
     journalService.saveBatch(jsonArray, TENANT_ID);
 
-    verify(journalRecordDao, times(3)).save(any(JournalRecord.class), eq(TENANT_ID));
+    verify(journalRecordDao, times(0)).save(any(JournalRecord.class), eq(TENANT_ID));
+    verify(journalRecordDao, times(1)).saveBatch(any(JsonArray.class), eq(TENANT_ID));
   }
 }